Move maxDirectMemory api to the java adaptor module
