+++ agent/agent-java-adaptor/agent-java-adaptor-java9/pom.xml
@@ -32,6 +32,9 @@
         <configuration>
           <source>9</source>
           <target>9</target>
+          <compilerArgs>
+            <arg>--add-exports=java.base/jdk.internal.misc=ALL-UNNAMED</arg>
+          </compilerArgs>
         </configuration>
       </plugin>
     </plugins>
