+++ agent/agent-observability/src/main/java/org/bithon/agent/observability/metric/collector/jvm/JvmMetricCollector.java
@@ -39,8 +39,6 @@ public class JvmMetricCollector {
     private CpuMetricCollector cpuMetricCollector;
 
     public void start() {
-        MemoryMetricCollector.initDirectMemoryCollector();
-
         cpuMetricCollector = new CpuMetricCollector();
 
         //
