--- agent/agent-observability/src/main/jdk8/org/bithon/agent/observability/metric/collector/jvm/MaxDirectMemoryCollectorJdk8.java
@@ -1,46 +1,0 @@
-/*
- *    Copyright 2020 bithon.org
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.bithon.agent.observability.metric.collector.jvm;
-
-import org.bithon.agent.instrumentation.expt.AgentException;
-
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-
-/**
- * JDK 8 implementation for collecting maximum direct memory using sun.misc.VM.
- * Since this class is compiled via JDK 17, we have to use reflection to access classes in JDK8 only.
- *
- * @author frank.chen021@outlook.com
- * @date 2025/08/30
- */
-class MaxDirectMemoryCollectorJdk8 implements DirectMemoryCollector.IMaxDirectMemoryGetter {
-
-    @Override
-    public long getMaxDirectMemory() {
-        try {
-            Class<?> vmClass = Class.forName("sun.misc.VM");
-            Method maxDirectMemoryMethod = vmClass.getDeclaredMethod("maxDirectMemory");
-            maxDirectMemoryMethod.setAccessible(true);
-            return (long) maxDirectMemoryMethod.invoke(null);
-        } catch (ClassNotFoundException | NoSuchMethodException | IllegalAccessException |
-                 InvocationTargetException e) {
-            throw new AgentException("sun.misc.VM.maxDirectMemory() not available", e);
-        }
-    }
-}
-
